feat: 只有购物车拥有者可以删除购物车中的商品

+++ app/Http/Controllers/CartItemController.php
@@ -28,6 +28,8 @@ class CartItemController extends Controller
 
     public function destroy(CartItem $cartItem)
     {
+        $this->authorize('delete', $cartItem);
+
         $cartItem->delete();
     }
 }
